outsourced Modeling from OLCModeler and DepModeler
